crawling

+++ services/swupdate/reprobuild/crawler.py
@@ -12,6 +12,10 @@ try:
 except:
     raise
 
+simulate = False
+simulateSuccess = False
+url_repro_build = 'https://tests.reproducible-builds.org'
+url_repro_testing = 'https://tests.reproducible-builds.org/debian/rb-pkg/testing/amd64/'
 
 packages_required = ['attr', 'base-files', 'base-passwd', 'debconf', 'debianutils', 'diffutils',
                      'dpkg', 'findutils', 'grep', 'gzip', 'init-system-helpers', 'libselinux', 'libsepol',
@@ -57,6 +61,7 @@ def docker(cmd, args=[]):
     cmds = ['docker'] + cmd.split(" ") + args
     if dlog:
         dlog.write(('\n\n--> %s\n\n' % '::'.join(cmds)).encode('utf-8'))
+        dlog.flush()
         proc = subprocess.Popen(cmds, stdout=dlog, stderr=dlog)
         proc.wait()
         return ""
@@ -82,12 +87,13 @@ def compile_bin(dname, did, dependencies, version, short_version, bina):
     comhash = ''
     wall_start_time = time.perf_counter()
     cpu_user_start, cpu_system_start = psutil.cpu_times().user, psutil.cpu_times().system
-    docker_exec(dname, "apt-get install -y " + ' '.join(dependencies))
-    full_pkg = name + '=' + version
-    docker_exec(dname, "apt-get source " + full_pkg)
-    docker_exec(dname, "apt-get build-dep -y --force-yes " + full_pkg)
-    src_dir = dir + '-' + short_version.partition('-')[0] + '/'
-    docker_exec(dname, "cd %s; dpkg-buildpackage -us -uc -tc" % src_dir)
+    if not simulate:
+        docker_exec(dname, "apt-get install -y " + ' '.join(dependencies))
+        full_pkg = name + '=' + version
+        docker_exec(dname, "apt-get source " + full_pkg)
+        docker_exec(dname, "apt-get build-dep -y --force-yes " + full_pkg)
+        src_dir = dir + '-' + short_version.partition('-')[0] + '/'
+        docker_exec(dname, "cd %s; dpkg-buildpackage -us -uc -tc -j4" % src_dir)
 
     cpu_user, cpu_system = psutil.cpu_times().user - cpu_user_start, psutil.cpu_times().system - cpu_system_start
     wall_time = time.perf_counter() - wall_start_time
@@ -117,11 +123,16 @@ def compile_bin(dname, did, dependencies, version, short_version, bina):
 
 # Find and add two Debian snapshots preceding the build time and one
 # following the build time
-def find_snapshots(btime, name):
+def find_snapshots(pkg, name):
+    page = urlopen(url_repro_testing + pkg + '.html').read()
+    soup = BeautifulSoup(page, 'html.parser')
+    timestr = soup.body.header.find('span', {'class': 'build-time'}).string.split()[1:3]
+    build_time = datetime.strptime(' '.join(timestr), '%Y-%m-%d %H:%M')
+
     snapbuf = []
     snapbuf_future = []
     for day in [-1,0,1]:
-        atime = btime + timedelta(days=day)
+        atime = build_time + timedelta(days=day)
         # print("Fetching", atime)
         snappage = urlopen(
             'http://snapshot.debian.org/archive/debian/?year=%s;month=%s' %
@@ -150,7 +161,8 @@ def find_snapshots(btime, name):
                 docker_exec(name, 'echo %s >> /etc/apt/sources.list' % line)
     docker_exec(name, "echo 'Acquire::Check-Valid-Until \"false\";' >> /etc/apt/apt.conf" )
     docker_exec(name, 'cat /etc/apt/sources.list')
-    docker_exec(name, 'apt-get update')
+    if not simulate:
+        docker_exec(name, 'apt-get update')
 
 # Save build results into csv file
 def save_results(yes, no, fail):
@@ -195,36 +207,26 @@ def get_packages(option):
 
 
 packages = get_packages(sys.argv[1])
-baseurl = 'https://tests.reproducible-builds.org'
-url = 'https://tests.reproducible-builds.org/debian/rb-pkg/testing/amd64/'
 hash_match, hash_differ, failed = [], [], []
 docker_build()
 
-for p in packages:
-    pkgstr = sys.argv[1]+'/'+p
-    # print("Reproducibly building package:", pkgstr)
-    page = urlopen(url + p + '.html').read()
-    soup = BeautifulSoup(page, 'html.parser')
-
-    # Retrieve build time
-    timestr = soup.body.header.find('span', {'class': 'build-time'}).string.split()[1:3]
-    build_time = datetime.strptime(' '.join(timestr), '%Y-%m-%d %H:%M')
-
-    dname = '-'.join(['repro', p, str(os.getpid())])
+def docker_run(pkg):
+    global dlog
+    dname = '-'.join(['repro', pkg, str(os.getpid())])
     did = docker('run --name=%s -d repro_build bash -c' % dname,["sleep 3600"])
-    dlog = open(p+'.log', 'wb')
+    dlog = open(pkg + '.log', 'wb')
     time.sleep(1)
     docker_exec(dname, "echo 193.62.202.30 snapshot.debian.org >> /etc/hosts")
-    find_snapshots(build_time, dname)
+    return dname, did
 
-    # Find all the dependencies
-    page = urlopen(baseurl + soup.body.header.find('a', {'title': 'Show: build info'})['href']).read()
+def find_dependencies(pkg):
+    page = urlopen(url_repro_testing + pkg + '.html').read()
     soup = BeautifulSoup(page, 'html.parser')
-
+    page = urlopen(url_repro_build + soup.body.header.find('a', {'title': 'Show: build info'})['href']).read()
+    soup = BeautifulSoup(page, 'html.parser')
     lines = str(soup).split('\n')
     shaflag = dflag = False
     version = name = sha = binary = dir = short_version = size = ''
-    first = True
     i = 0
     dependencies = []
     for line in lines:  # Extracting describing data
@@ -238,10 +240,10 @@ for p in packages:
                     else:
                         version = short_version = words[1]
                 elif words[0] == 'Binary:':  # To know binary name for the verified package
-                    if p in words:
-                        name = p
-                    elif (p + '1') in words:
-                        name = p + '1'
+                    if pkg in words:
+                        name = pkg
+                    elif (pkg + '1') in words:
+                        name = pkg + '1'
                     else:
                         name = words[1]
                 elif words[0] == 'Source:':  # To know a name of the folder where to compile a package
@@ -261,31 +263,51 @@ for p in packages:
             i += 1
             dependencies.append(parse_dpnd(line))
 
-    computed_hash, wtime, utime, stime = \
+    return dependencies, version, name, sha, binary, dir, short_version, size
+
+
+for pkg in packages:
+    pkgstr = sys.argv[1]+'/' + pkg
+    # print("Reproducibly building package:", pkgstr)
+
+    dname, did = docker_run(pkg)
+
+    # Get all snapshots
+    find_snapshots(pkg, dname)
+
+    # Find all the dependencies
+    dependencies, version, name, sha, binary, dir, short_version, size = \
+        find_dependencies(pkg)
+
+    if not simulate:
+        computed_hash, wtime, utime, stime = \
         compile_bin(dname, did, dependencies, version, short_version, binary)
-    # computed_hash, wtime, utime, stime = sha, 1.5, 2.5, 3.5
-    # print("Got", computed_hash, wtime, utime, stime)
+    else:
+        computed_hash, wtime, utime, stime = "", 1.5, 2.5, 3.5
+        if simulateSuccess:
+            computed_hash = sha
+        print("Got", computed_hash, wtime, utime, stime)
 
     if sha != '' and computed_hash == sha:
-        hash_match.append((p, binary, size, wtime, utime, stime, 'y'))
+        hash_match.append((pkg, binary, size, wtime, utime, stime, 'y'))
         print("Hashes match for %s: wall(%f) user(%f) system(%f)" %
               (pkgstr, wtime, utime, stime))
     else:
         if computed_hash == '':
-            failed.append((p, binary, size, wtime, utime, stime, 'f'))
+            failed.append((pkg, binary, size, wtime, utime, stime, 'f'))
             print('Fail in the build process')
         else:
-            hash_differ.append((p, binary, size, wtime, utime, stime, 'n'))
+            hash_differ.append((pkg, binary, size, wtime, utime, stime, 'n'))
             print("Hashes differ for", pkgstr, computed_hash, "and", sha)
 
 save_results(hash_match, hash_differ, failed)
 
-print('\nBuilt packages with matching hash: ', hash_match)
+print('Built packages with matching hash: ', hash_match)
 print('Failed to build:', failed)
 print('Built packages with differed hash:', hash_differ)
 
 if sys.argv[1] == 'cli':
-    if len(hash_match) > 1:
+    if len(hash_match) > 0:
         print(sys.argv[2], hash_match[0])
         print("Success", wtime, utime, stime)
     else:
